feat: refactor and add roads
