+++ src/domain/entities/map/MapColorsType.ts
@@ -1,0 +1,7 @@
+export interface MapColorsType {
+  activeCountries: string;
+  inactiveCountries: string;
+  ocean: string;
+  outline: string;
+  roads: string;
+}
