+++ tailwind.config.js
@@ -72,16 +72,6 @@ module.exports = {
   },
   darkMode: 'class',
   plugins: [
-    function ({ addBase, theme }) {
-      addBase({
-        ':root': {
-          '--color-hover': theme('colors.hover'),
-          '--color-background': theme('colors.background'),
-          '--color-active-countries': theme('colors.activeCountries'),
-          '--color-inactive-countries': theme('colors.inactiveCountries'),
-        },
-      });
-    },
     nextui({
       themes: {
         light: {
@@ -109,8 +99,6 @@ module.exports = {
             chatbotUserMsg: '#E6F1FE',
             chatbotDivider: '#292d32',
             surfaceGrey: '#B0B0B0',
-            activeCountries: '#82bce0',
-            inactiveCountries: '#a7b3ba',
           },
         },
         dark: {
@@ -138,8 +126,6 @@ module.exports = {
             chatbotUserMsg: '#26262A',
             chatbotDivider: '#556372',
             surfaceGrey: '#444444',
-            activeCountries: '#115884',
-            inactiveCountries: '#85929b',
           },
         },
       },
